core

Toolips.AbstractRoute — Type

abstract type AbstractRoute

Abstract Routes are what connect incoming connections to functions. Each route must have two fields, path, and page. Path needs to be a String, but that is about it.

Consistencies

type::T where T == Vector{Symbol} || T == Symbol

Toolips.Servable — Type

abstract type Servable

Servables can be written to a Connection via their f() function and the interface. They can also be indexed with strings or symbols to change properties

Consistencies

f::Function - Function whose output to be written to http. Must take a single

positonal argument of type ::Connection or ::AbstractConnection

Toolips.AbstractConnection — Type

abstract type AbstractConnection

Connections are passed through function routes and can have Servables written to it.

Consistencies

routes::Dict - A {String, Function} dictionary that the server references to

direct incoming connections.

http::Any - Usually an HTTP.Stream, however can be anything that is binded to

the Base.write method, or the Toolips.write! method.

extensions::Dict - A {Symbol, ServerExtension} dictionary that can be used to

access ServerExtensions.

Toolips.ServerExtension — Type

abstract type ServerExtension

Server extensions are loaded into the server on startup, and can have a few different abilities according to their type field's value. This value can be either a Symbol or a Vector of Symbols.

Consistencies

type::T where T == Vector{Symbol} || T == Symbol. The type can be :routing,

:func, :connection, or any combination inside of a Vector{Symbol}.

:routing extensions are called once at server creation, and must have

the field f(r::Vector{AbstractRoute}, e::Vector{ServerExtension}).

:func extensions are called each time the server is routed, and must have

the field f(c::AbstractConnection).

:connection extensions are passed inside of the Connection.

Toolips.ToolipsServer — Type

abstract type ToolipsServer

ToolipsServers are returned whenever the ServerTemplate.start() field is called. If you are running your server as a module, it should be noted that commonly a global start() method is used and returns this server, and dev is where this module is loaded, served, and revised.

Consistencies

routes::Vector{AbstractRoute} - The server's routes.

extensions::Vector{Route} - The server's currently loaded extensions.

server::Any - The server, whatever type it may be...

Toolips.CoreException — Type

abstract type CoreException

Core Exceptions are thrown whenever a random Core error happens.

Consistencies

type::T where T == Vector{Symbol} || T == Symbol

connections

Toolips.SpoofStream — Type

SpoofStream

text::String

The SpoofStream allows us to fake a connection by building a SpoofConnection which will write to the SpoofStream.text field whenever write! is called. This is useful for testing, or just writing servables into a string.

example

stream = SpoofStream()
write(stream, "hello!")
println(stream.text)

    hello!
conn = SpoofConnection()
servab = Component()
write!(conn, servab)

field info

text::String - The text written to the stream.

constructors

SpoofStream()

Toolips.Connection — Type

Connection <: AbstractConnection

routes::Dict

http::HTTP.Stream

extensions::Dict

The connection type is passed into route functions and pages as an argument. This is both for functions, as well as Servable.f() methods. This constructor should not be called directly. Instead, it is called by the server and passed through the function pipeline. Indexing a Connection will return the extension named with that symbol.

example

                  #  v The Connection
home = route("/") do c::Connection
    c[Logger].log(1, "We can index extensions by type or symbol")
    c[:logger].log(1, "see?")
    c.routes["/"] = c::Connection -> write!(c, "rerouting!")
    httpstream = c.http
    write!(c, "Hello world!")
    myheading::Component = h("myheading", 1, text = "Whoa!")
    write!(c, myheading)
end

field info

routes::Dict - A dictionary of routes where the keys

are the routed URL and the values are the functions to those keys.

http::HTTP.Stream - The stream for this current peer's connection.

extensions::Dict - A dictionary of extensions to load with the

name to reference as keys and the extension as the pair.

constructors

Connection(routes::Dict, http::HTTP.Stream, extensions::Dict)

Toolips.SpoofConnection — Type

SpoofConnection <: AbstractConnection

routes::Dict

http::SpoofStream

extensions::Dict -

Builds a fake connection with a SpoofStream. Useful if you want to write a Servable without a server.

example

fakec = SpoofConnection()
servable = Component()
# write!(::AbstractConnection, ::Servable):
write!(fakec, servable)

field info

routes::Dict - A dictionary of routes, usually left empty.

http::SpoofStream - A fake http stream that instead writes output to a string.

extensions::Dict - A dictionary of extensions, usually empty.

constructors

SpoofStream(r::Dict, http::SpoofStream, extensions::Dict)

SpoofStream()

Base.getindex — Method

Core

getindex(c::AbstractConnection, s::Symbol) -> ::ServerExtension

Indexes the extensions in c.

example

route("/") do c::Connection
    c[:Logger].log("hi")
end

Base.getindex — Method

Core

getindex(c::AbstractConnection, t::Type) -> ::ServerExtension

Indexes the extensions in c by type.

example

route("/") do c::Connection
    c[Logger].log("hi")
end

Base.setindex! — Method

Core

setindex!(c::AbstractConnection, f::Function, s::String) -> _

Sets the route path s to serve at the function f.

example

c["/"] = c -> write!(c, "hello")

Base.setindex! — Method

Core

setindex!(c::AbstractConnection, f::Function, s::String) -> _

Sets the route path s to serve at the function f.

example

c["/"] = c -> write!(c, "hello")

Toolips.write! — Function

Interface

write!(c::AbstractConnection, s::Servable) -> _

Writes a Servable's return to a Connection's stream. This is usually used in a routing function or a route where ::Connection is provided as an argument.

example

serv = p("mycomp", text = "hello")

rt = route("/") do c::Connection
    write!(c, serv)
end

Interface

write!(c::AbstractConnection, s::Vector{Servable}) -> _

Writes all servables in s to c.

example

c = AbstractComponent()
c2 = AbstractComponent()
comps = components(c, c2)
    Vector{Servable}(AbstractComponent(), AbstractComponent())

write!(c, comps)

Interface

write!(c::AbstractConnection, s::Servable ...) -> _

Writes Servables as Vector{Servable}

example

write!(c, p("mycomp", text = "hello!"), p("othercomp", text = "hi!"))

Interface

write!(c::AbstractConnection, s::Vector{AbstractComponent}) -> _

A catch-all for when Vectors are accidentally stored as Vector{Any}.

example

write!(c, ["hello", p("mycomp", text = "hello!")])

Interface

write!(::AbstractConnection, ::Any) -> _

Attempts to write any type to the Connection's stream.

example

d = 50
write!(c, d)

Toolips.getargs — Function

Core

getargs(c::AbstractConnection) -> ::Dict{Symbol, Any}

The getargs method returns arguments from the HTTP target (GET requests.) Returns a Dict with the argument keys as Symbols.

example

route("/") do c
    args = getargs(c)
    args[:message]
        "welcome to toolips ! :)"
end

Toolips.getarg — Function

Core

getarg(c::AbstractConnection, s::Symbol) -> ::Any

Returns the requested argument from the target.

example

getarg(c, :x)
    50

Core

getarg(c::AbstractConnection, s::Symbol, t::Type) -> ::Vector

This method is the same as getargs(::HTTP.Stream, ::Symbol), however types are parsed as type T(). Note that "Cannot convert..." errors are possible with this method.

example

getarg(c, :x, Int64)
    50

Toolips.getip — Function

Core

getip(c::AbstractConnection) -> ::String

Returns the IP that is connected via the connection c.

example

getip(c)
"127.0.0.2"

Toolips.getpost — Function

Core

getpost(c::AbstractConnection) -> ::String

Returns the POST body of c.

example

getpost(c)
"hello, this is a post request"

Toolips.download! — Function

Core

download!(c::AbstractConnection, uri::String) -> _

Downloads a file to a given Connection's computer.

example

download!(c, "files/mytext.txt")

Toolips.navigate! — Function

Core

navigate!(::AbstractConnection, ::String) -> _

Routes a connected stream to a given URL.

example

navigate!(c, "https://github.com/ChifiSource/Toolips.jl")

Base.push! — Method

Interface

push!(c::AbstractConnection, data::Any) -> _

A "catch-all" for pushing data to a stream. Produces a full response with data as the body.

example

Toolips.startread! — Method

Interface

startread!(::AbstractConnection) -> _

Resets the seek on the Connection. This function is only meant to be used on post bodies.

example

post = getpost(c)
    "hello"
post = getpost(c)
    ""
startread!(c)
post = getpost(c)
    "hello"

Toolips.route! — Method

Interface

route!(c::AbstractConnection, route::Route) -> _

Modifies the route on the Connection.

example

route("/") do c::Connection
    r = route("/") do c::Connection
        write!(c, "hello")
    end
    route!(c, r)
end

Toolips.route! — Method

Interface

route!(::Function, ::AbstractConnection, ::String) -> _

Routes a given String to the Function.

example

route("/") do c
    route!(c, "/") do c
        println("tacos")
    end
end

Toolips.unroute! — Method

Interface

unroute!(::AbstractConnection, ::String) -> _

Removes the route with the key equivalent to the String.

example

# One request will kill this route:
route("/") do c::Connection
    unroute!(c, "/")
end

Toolips.routes — Method

Interface

routes(c::Connection) -> ::Dict{String, Function}

Returns the server's routes.

example

route("/") do c::Connection
    routes(c)
end

Toolips.extensions — Method

Interface

extensions(c::Connection) -> ::Dict{Symbol, ServerExtension}

Returns the server's extensions.

example

route("/") do c::Connection
    extensions(c)
end

Toolips.has_extension — Function

Core

has_extension(c::AbstractConnection, t::Type) -> ::Bool

Checks if c.extensions has an extension of type t.

example

if has_extension(c, Logger)
    c[:Logger].log("it has a logger, I think.")
end

Internals

has_extension(d::Dict, t::Type) -> ::Bool

Checks if d has an extension of type t.

example

if has_extension(d, Logger)
    d[:Logger].log("it has a logger, I think.")
end

routes

Toolips.Route — Type

Route

path::String - The path to route to the function, e.g. "/".

page::Function - The function to route the path to.

A route is added to a ToolipsServer using either its constructor, or the ToolipsServer.add(::Route) method. Each route calls a function. The Route type is commonly constructed using the do syntax with the route(::Function, ::String) method.

example

# Constructors
route = Route("/", p(text = "hello"))

function example(c::Connection)
    write!(c, "hello")
end

route = Route("/", example)

# method
route = route("/") do c
    write!(c, "Hello world!")
    write!(c, p(text = "hello"))
    # we can also use extensions!
    c[:logger].log("hello world!")
end

constructors

Route(path::String, f::Function)

Toolips.route — Function

Interface

route(f::Function, r::String) -> ::Route

Creates a route from the Function. The function should take a Connection or AbstractConnection as a single positional argument.

example

route("/") do c::Connection

end

Interface

route(r::String, f::Function) -> ::Route

Creates a route from the Function. The function should take a Connection or AbstractConnection as a single positional argument.

example

function example(c::Connection)
    write!(c, h("myh", 1, text = "hello!"))
end
r = route("/", example)

Toolips.routes — Function

Interface

routes(c::Connection) -> ::Dict{String, Function}

Returns the server's routes.

example

route("/") do c::Connection
    routes(c)
end

Interface

routes(::Route ...) -> ::Vector{Route}

Turns routes provided as arguments into a Vector{Route} with indexable routes. This is useful because this is the type that the ServerTemplate constructor likes. This function is also used as a "getter" for ToolipsServers and Connections, see ?(routes(::ToolipsServer)) & ?(routes(::AbstractConnection))

example

r1 = route("/") do c::Connection
    write!(c, "pickles")
end
r2 = route("/pickles") do c::Connection
    write!(c, "also pickles")
end
rts = routes(r1, r2)

Interface

routes(ws::ToolipsServer) -> ::Dict{String, Function}

Returns the server's routes.

example

ws = MyProject.start()
routes(ws)
    "/" => home
    "404" => fourohfour

Base.setindex! — Method

Interface

setindex!(rs::Vector{AbstractRoute}, f::Function, s::String)

Sets a given route in a Vector to a function.

example

function example(c::Connection)
    write!(c, h("myh", 1, text = "hello!"))
end
r = route("/", example)
rts = [r]
rts["/"] = (c::Connection) -> write!(c, p(text = "no longer a heading"))

servers

Toolips.WebServer — Type

WebServer <: ToolipsServer

host::String

routes::Dict

extensions::Dict

server::Any

add::Function

remove::Function

start::Function

A web-server is given as a return from a ServerTemplate whenever ServerTemplate.start() is ran. It can be rerouted with route! and indexed similarly to the Connection, with Symbols representing extensions and Strings representing routes.

example

st = ServerTemplate()
ws = st.start()
routes(ws)
...
extensions(ws)
...
route!(ws, "/") do c::Connection
    write!(c, "hello")
end

Toolips.ServerTemplate — Type

ServerTemplate

host::String

port::Integer

routes::Vector{AbstractRoute}

extensions::Vector{ServerExtension}

remove::Function

add::Function

start::Function

The ServerTemplate is used to configure a server before running. These are commonly used for reproducibility, especially when it comes to making servers from extensions

example

home(c::Connection) = begin
    write!(c, p(text = "hello world!"))
end

st = ServerTemplate("127.0.0.1", 8000, routes = [Route("/", home)])

webserver = ServerTemplate.start()

field info

ip::String - IP the server should serve to.

port::Integer - Port to listen on.

routes::Vector{AbstractRoute} - A vector of routes to provide to the server

extensions::Vector{ServerExtension} - A vector of extensions to load into

the server.

remove(::Int64)::Function - Removes routes by index.

remove(::String)::Function - Removes routes by name.

remove(::Symbol)::Function - Removes extension by Symbol representing

type, e.g. :Logger

add(::Route ...)::Function - Adds the routes to the server.

add(::ServerExtension ...)::Function - Adds the extensions to the server.

start()::Function - Starts the server.

constructors

ServerTemplate(ip::String = "127.0.0.1", port::Int64 = 8001, routes::Vector{AbstractRoute} = Vector{AbstractRoute}()); extensions::Vector{ServerExtension} = [Logger()] connection::Type)

Base.getindex — Method

Interface

getindex(ws::ToolipsServer, s::Symbol) -> ::ServerExtension

Indexes the extensions in ws.

example

ws = MyProject.start()
ws[:Logger].log("hi")

Toolips.routes — Method

Interface

routes(ws::ToolipsServer) -> ::Dict{String, Function}

Returns the server's routes.

example

ws = MyProject.start()
routes(ws)
    "/" => home
    "404" => fourohfour

Toolips.extensions — Method

Interface

extensions(ws::ToolipsServer) -> ::Dict{Symbol, ServerExtension}

Returns the server's extensions.

example

ws = MyProject.start()
extensions(ws)
    Logger(blah blah blah)

Toolips.kill! — Method

Interface

kill!(ws::ToolipsServer) -> _

Closes the web-server.

example

ws = MyProject.start()
kill!(ws)

Toolips.kill! — Method

Interface

kill!(ws::ToolipsServer) -> _

Closes the web-server.

example

ws = MyProject.start()
kill!(ws)

Toolips.route! — Method

Interface

route!(f::Function, ws::ToolipsServer, r::String) -> _

Reroutes a server's route r to function f.

example

ws = MyProject.start()
route!(ws, "/") do c
    c[:Logger].log("rerouted!")
end

Toolips.route! — Method

Interface

route!(ws::ToolipsServer, r::String, f::Function) -> _

Reroutes a server's route r to function f.

example

ws = MyProject.start()

function myf(c::Connection)
    write!(c, "pasta")
end
route!(ws, "/", myf)

Toolips.route! — Method

Interface

route!(ws::ToolipsServer, r::Route) -> _

Reroutes a server's route r.

example

ws = MyProject.start()
r = route("/") do c

end
route!(ws, r)

requests

Base.get — Method

Core

get(url::String) -> ::String

Quick binding for an HTTP GET request.

example

body = get("/")
    "hi"

Toolips.post — Function

Core

post(url::String, body::String) -> ::String

Quick binding for an HTTP POST request.

example

response = post("/")
    "my response"