Interface

Toolips.AbstractComponent — Type

abstract type AbstractComponent <: Servable

Components are html elements.

Consistencies

properties::Dict{Any, Any}

extras::Vector{Servable}

name::String

Servable Consistencies

Servables can be written to a Connection via thier f() function and the
interface. They can also be indexed with strings or symbols to change properties
##### Consistencies
- f::Function - Function whose output to be written to http. Must take a single
positonal argument of type ::Connection or ::AbstractConnection

Toolips.StyleComponent — Type

abstract type StyleComponent <: Servable

Not much different from a normal AbstractComponent, simply an abstract type step for the interface to separate working with Animations and Styles.

AbstractComponent Consistencies

properties::Dict{Any, Any}

extras::Vector{Servable}

name::String

Servable Consistencies

Servables can be written to a Connection via thier f() function and the
interface. They can also be indexed with strings or symbols to change properties
##### Consistencies
- f::Function - Function whose output to be written to http. Must take a single
positonal argument of type ::Connection or ::AbstractConnection

file

Toolips.File — Type

File <: Servable

dir::String f::Function Serves a file into a Connection.

example

f = File("hello.txt")
r = route("/") do c
    write!(c, f)
end

field info

dir::String - The directory of a file to serve.

f::Function - Function whose output to be written to http().

constructors

File(dir::String)

components

Toolips.Component — Type

Component{tag} <: AbstractComponent <: Servable

name::String

f::Function

tag::String

properties::Dict

A component is a standard servable which is used to represent HTML tag structures. Indexing a Component with a Symbol or a String will return or set a Component's property to that index. The two special indexes are :children and :text. :text will change the inner content of the Component and :children is where components that will be written inside the Component go. You can add to these with push!(c::Servable, c2::Servable)

example

using Toolips

image_style = Style("example")
image_anim = Animation("img_anim")
image_anim[:from] = "opacity" => "0%"
image_anim[:to] = "opacity" => "100%"
animate!(image_style)

r = route("/") do c::AbstractConnection
    newimage = img("newimage", src = "/logo.png")
    style!(newimage, image_style)
    write!(c, newimage)
end

field info

name::String - The name field is the way that a component is denoted in code.

f::Function - The function that gets called with the Connection as an

argument.

properties::Dict - A dictionary of symbols and values.

constructors

Component(name::String = "", tag::String = "", properties::Dict = Dict())

Component(name::String, tag::String, props::Base.Pairs)

Toolips.Style — Type

Style

name::String

f::Function

properties::Dict{Any, Any}

extras::Vector{Servable}

Creates a style from attributes, can style a Component using the style! method. Names should be consistent with CSS names. For example, a default h1 style would be named "h1". A heading style for a specific class should be "h1.myheading"

example

style = Style("p.mystyle", color = "blue")
style["opacity"] = "50%"
comp = Component()
style!(comp, style)

field info

name::String - The name of the style. Should be consistent with CSS naming.

f::Function - The function f, called by write! when writing to a Connection.

properties::Dict{Any, Any} - A dict of style attributes.

extras::String - Extra components to be written along with the style. Usually

this is an animation.

constructors

Style(name::String; props ...)

Toolips.Animation — Type

Animation

name::String

properties::Dict

f::Function

delay::Float64

length::Float64

iterations::Integer

An animation can be used to animate Styles with the animate! method. Animating is done by indexing by either percentage, or symbols, such as from and to.

example

anim = Animation("myanim")
anim[:from] = "opacity" => "0%"
anim[:to] = "opacity" => "100%"
style = Style("example")
animate!(style, anim)

field info

name::String - The name of the animation.

properties::Dict - The properties that have been pushed so far.

f::Function - The function called when writing to a Connection.

delay::Float64 - The delay before the animation begins.

length::Float64 - The amount of time the animation should play.

iterations::Integer - The number of times the animation should repeat. When

set to 0 the animation will loop indefinitely.

constructors

Animation(name::String = "animation", delay::Float64 = 0.0, length::Float64 = 5.2, iterations::Integer = 1)

Base.getindex — Method

Interface

getindex(s::AbstractComponent, symb::Symbol) -> ::Any

Returns a property value by symbol or name.

example

c = p("hello", text = "Hello world")
c[:text]
    "Hello world!"

c["opacity"] = "50%"
c["opacity"]
    "50%"

Base.getindex — Method

Interface

getindex(::Servable, ::String) -> ::Any

Returns a property value by string or name.

example

c = p("hello", text = "Hello world")
c[:text]
    "Hello world!"

c["opacity"] = "50%"
c["opacity"]
    "50%"

Base.setindex! — Method

Interface

setindex!(s::Servable, a::Any, symb::Symbol) -> _

Sets the property represented by the symbol to the provided value.

example

c = p("world")
c[:text] = "hello world!"

Missing docstring.

Missing docstring for setindex!(::Servable, ::Any, ::String). Check Documenter's build log for details.

Toolips.img — Function

img(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the img Component with the key-word arguments provided in args as properties.

example

image = img("mylogo", src = "assets/logo.png")
write!(c, image)

Toolips.link — Function

link(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the link Component with the key-word arguments provided in args as properties.

example

mylink = link("mylink", href = "http://toolips.app")
write!(c, mylink)

Toolips.meta — Function

meta(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the meta Component with the key-word arguments provided in args as properties.

example

metainfo = meta("metainfo", rel = "meta-description", text = "hello")
write!(c, metainfo)

Toolips.input — Function

input(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the input Component with the key-word arguments provided in args as properties.

example

element = input("mylogo")
write!(c, element)

Toolips.a — Function

a(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the a Component with the key-word arguments provided in args as properties.

example

element = a("mylogo")
write!(c, element)

Toolips.p — Function

p(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the p Component with the key-word arguments provided in args as properties.

example

p1 = input("mylogo")
write!(c, p)

Toolips.h — Function

h(name::String, level::Integer = 1, , p::Pair{String, Any} ...; keys ...) -> ::Component

Returns the h Component with the key-word arguments provided in args as properties.

example

h1 = h("heading1", 1, text = "hello world!")
write!(c, h1)

Toolips.h1 — Function

h1(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the h1 Component with the key-word arguments provided in args as properties.

example

h1 = h("heading1", 1)
write!(c, h1)

Toolips.ul — Function

ul(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the ul Component with the key-word arguments provided in args as properties.

example

ul1 = ul("mylogo")
write!(c, ul)

Toolips.li — Function

li(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the li Component with the key-word arguments provided in args as properties.

example

li1 = li("mylogo")
write!(c, li)

Base.div — Function

div(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the div Component with the key-word arguments provided in args as properties.

example

divider1 = div"mylogo")
write!(c, divider1)

Toolips.divider — Function

divider(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the div Component with the key-word arguments provided in args as properties.

example

divider1 = divider("mylogo")
write!(c, divider1)

Toolips.br — Function

br(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the br Component with the key-word arguments provided in args as properties.

example

comp = br("newcomp")
write!(c, comp)

Toolips.i — Function

i(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the i Component with the key-word arguments provided in args as properties.

example

comp = i("newcomp")
write!(c, comp)

Toolips.title — Function

title(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the title Component with the key-word arguments provided in args as properties.

example

comp = title("newcomp")
write!(c, comp)

Toolips.span — Function

span(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the span Component with the key-word arguments provided in args as properties.

example

comp = span("newcomp")
write!(c, comp)

Toolips.iframe — Function

iframe(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the iframe Component with the key-word arguments provided in args as properties.

example

comp = iframe("newcomp")
write!(c, comp)

Toolips.svg — Function

svg(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the svg Component with the key-word arguments provided in args as properties.

example

comp = svg("newcomp")
write!(c, comp)

Toolips.element — Function

element(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the element Component with the key-word arguments provided in args as properties.

example

comp = element("newcomp")
write!(c, comp)

Toolips.label — Function

label(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the label Component with the key-word arguments provided in args as properties.

example

lbl = label("mylogo", src = "assets/logo.png")
write!(c, lbl)

Toolips.script — Function

script(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the script Component with the key-word arguments provided in args as properties.

example

comp = script("newcomp")
write!(c, comp)

Toolips.nav — Function

nav(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the nav Component with the key-word arguments provided in args as properties.

example

comp = nav("newcomp")
write!(c, comp)

Toolips.button — Function

button(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the button Component with the key-word arguments provided in args as properties.

example

comp = button("newcomp")
write!(c, comp)

Toolips.form — Function

form(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = form("newcomp")
write!(c, comp)

Toolips.section — Function

section(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = section("newcomp")
write!(c, comp)

Toolips.body — Function

body(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = body("newcomp")
write!(c, comp)

Toolips.header — Function

header(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = header("newcomp")
write!(c, comp)

Toolips.footer — Function

footer(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = footer("newcomp")
write!(c, comp)

Toolips.b — Function

footer(name::String, p::Pair{String, Any} ...; args ...) -> ::Component

Returns the form Component with the key-word arguments provided in args as properties.

example

comp = footer("newcomp")
write!(c, comp)

Toolips.DOCTYPE — Function

DOCTYPE() -> ::String

DOCTYPE occassionally needs to be written to the top of files to make HTML render properly.

example

write!(c, DOCTYPE())

Toolips.properties! — Function

Interface

properties!(c::Servable, s::Servable) -> _

Copies properties from s,properties into c.properties.

example

comp = AbstractComponent()
othercomp = AbstractComponent()
othercomp["opacity"] = "100%"
properties!(comp, othercomp)

comp["opacity"]
        100%

Toolips.getproperties — Function

Interface

getproperties(c::AbstractComponent) -> ::Dict

Returns a Dict of properties inside of c.

example

props = properties(c)

Toolips.children — Function

Interface

children(c::AbstractComponent) -> ::Vector{Servable}

Returns Vector{Servable} of children inside of c.

example

children(c)

Base.get — Method

Interface

get(c::AbstractComponent, s::String) -> ::Servable

Returns the child with name s from c's children.

example

comp = p("myp", text = "hello")
c = div("c", align = "center")
push!(c, comp)

get(c, "myp")
Component{:p}(...)

Toolips.copy — Function

Interface

copy(c::AbstractComponent) -> ::AbstractComponent

copies c.

example

c = p("myp")
t = copy!(c)

Interface

copy(c::Style) -> ::Style

copies c.

example

c = Style("div", "background-color" => "blue")
t = copy!(c)

Interface

copy(c::Animation) -> ::Animation

copies c.

example

c = Style("div", "background-color" => "blue")
t = copy!(c)

Toolips.has_children — Function

Interface

has_children(c::AbstractComponent) -> ::Bool

Returns true if the given component has children.

example

c = AbstractComponent()
otherc = AbstractComponent()
push!(c, otherc)

has_children(c)
    true
has_children(otherc)
    false

Base.push! — Method

Interface

push!(s::AbstractComponent, d::AbstractComponent ...) -> ::AbstractComponent

Adds the child or children d to s.properties[:children]

example

c = AbstractComponent()
otherc = AbstractComponent()
push!(c, otherc)

Toolips.style! — Function

Interface

style!(c::Servable, s::Style) -> _

Applies the style to a servable.

example

serv = p("wow")
mystyle = Style("mystyle", color = "lightblue")
style!(serv, mystyle)

Interface

style!(c::Servable, s::Pair ...) -> _

Applies the style pairs to the servable's "style" property.

example

mycomp = p("mycomp")
style!(mycomp, "background-color" => "lightblue", "color" => "white")

Interface

style!(c::Servable, s::Vector{Pair}) -> _

Applies the style pairs to the servable's "style" property.

example

mycomp = p("mycomp")
style!(mycomp, ["background-color" => "lightblue", "color" => "white"])

Interface

style!(::Style, ::Style) -> _

Copies the properties from the second style into the first style.

example

style1 = Style("firsts")
style2 = Style("seconds")
style1["color"] = "orange"
style!(style2, style1)

style2["color"]
    "orange"

Session Interface

style!(cm::Modifier, s::Servable, style::Style) -> _

Changes the style class of s to the style p. Note – styles must be already written to the Connection prior.

example

function home(c::Connection)
    mystyle = Style("newclass", "background-color" => "blue")
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, mybutton, mystyle)
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, name::String, sname::String) -> _

Changes the style class of a Servable by name to the style p by name. Note – styles must be already written to the Connection prior.

example

function home(c::Connection)
    mystyle = Style("newclass", "background-color" => "blue")
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, mybutton, "newclass") #<- name of mystyle
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, s::Servable, p::Pair{String, String}) -> _

Styles the Servable s with the properties and values in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, mybutton, "background-color" => "lightblue", "color" => "white")
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, s::String, p::Pair{String, String}) -> _

Styles the Servable s by name with the properties and values in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, "mybutton", "background-color" => "lightblue", "color" => "white")
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, s::Servable, p::Pair) -> _

Styles the Servable s with the property and value in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, mybutton, "background-color" => "lightblue")
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, name::String, p::Pair) -> _

Styles a Servable by name with the property and value in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, "mybutton", "background-color" => "lightblue")
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, name::String, p::Vector{Pair{String, String}}) -> _

Styles a Servable by name with the properties and values in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, mybutton, ["background-color" => "lightblue"])
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

Session Interface

style!(cm::Modifier, name::String, p::Vector{Pair{String, String}}) -> _

Styles a Servable by name with the properties and values in p.

example

function home(c::Connection)
    mybutton = button("mybutton", text = "click to change alignment")
    mydiv = div("mydiv")
    on(c, mybutton, "click") do cm::Modifier
        style!(cm, "mybutton", "background-color" => "lightblue")
    end
    write!(c, mybutton)
    write!(c, mydiv)
end

extensions

Toolips.Logger — Type

Logger <: ServerExtension

type::Symbol

out::String

levels::Dict

log::Function

prefix::String

timeformat::String

writeat::Int64 -

A Logger logs information with different levels. Holds the function log(), connected to the function _log(). Methods.

example

logger = Logger()
st = ServerTemplate(extensions = [Logger()])
r = route("/") do c::Connection
    write!(c, "hello world!")
    c[:Logger].log("Hello world delivered, mission accomplished.")
end
st.add(r)
st.start()

field info

type::Symbol - The type of server extension – in this case, Connection.

out::String - Logfile output directory.

log(level::Int64, message::String) - Logs the message at the provided level.

log(message::String) - Logs the message at level 1.

log(c::Connection, message::String) - Logs to level one and to JavaScript

console.

levels::Dict - A {Any, Crayon} dict that contains all of the crayons for the

logger. Also contains two special crayons under the keys :timecrayon and :messagecrayon

prefix::String - The prefix to write before the message.

timeformat::String - A string representing DT format, must be able to be

passed through the datetime_str macro from Dates.

writeat::Int64 - The log level to write to out at.

constructors

Logger(levels::Dict{level_count::Int64 => crayon::Crayons.Crayon}; out::String = pwd() * "logs/log.txt") Logger(; out::String = pwd() * "/logs/log.txt")

Toolips.Files — Type

Files <: ServerExtension

type::Symbol

directory::String

f::Function -

Writes all files in directory to their own routes in the server.

field info

type::Symbol - The type of extension. There are three different selections

you can choose from. :connection :routing :func. A :connection extension will be provided in Connection.extensions. A :routing function is passed a Dict of routes as an argument. The last is a function argument, which is just a specific function to run from the top-end to the server.

directory::String - The directory to route the files from.

f::Function - The function f() called with a Connection.

constructors

Files(dir::String)